backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
PyPDF2==3.0.1
python-dotenv==1.0.1
google-generativeai==0.8.3
